+++ src/main/java/shop/mtcoding/miniproject2/controller/company/CompanyResumeController.java
@@ -46,6 +46,7 @@ import shop.mtcoding.miniproject2.model.SkillFilterRepository;
 import shop.mtcoding.miniproject2.model.SkillRepository;
 import shop.mtcoding.miniproject2.model.User;
 import shop.mtcoding.miniproject2.model.UserRepository;
+import shop.mtcoding.miniproject2.service.CompanyService;
 
 @RequestMapping("/company")
 @RequiredArgsConstructor
@@ -61,6 +62,7 @@ public class CompanyResumeController {
     private final PersonRepository personRepository;
     private final CompanyScrapRepository companyScrapRepository;
     private final SkillFilterRepository skillFilterRepository;
+    private final CompanyService companyService;
 
     @GetMapping("/resumes")
     public ResponseEntity<?> resume() {
@@ -131,6 +133,7 @@ public class CompanyResumeController {
         return new ResponseEntity<>(new ResponseDto<>(1, "", null), HttpStatus.OK);
     }
 
+    // 공고 + 스킬 찾기
     @GetMapping("/recommend")
     public ResponseEntity<?> recommend() {
         User principal = (User) session.getAttribute("principal");
@@ -139,73 +142,9 @@ public class CompanyResumeController {
         }
 
         // 공고 + 스킬 찾기
-        List<postIdAndSkillsDto> postAndSkillsList = postRepository.findPostIdAndSkills(principal.getCInfoId());
 
-        List<ResumeWithPostInfoRecommendDto> resumeAndPostInfo = new ArrayList<>();
-        for (postIdAndSkillsDto p : postAndSkillsList) {
-            String[] skills = p.getSkills().split(",");
-            List<SkillFilter> sFilters = new ArrayList<>();
-            for (String skill : skills) {
-                List<SkillFilter> s = skillFilterRepository.findSkillNameForCompany(skill);
-                sFilters.addAll(s);
-            }
-
-            // resume id로 count
-            HashMap<Integer, Integer> resumeIdAndCount = new HashMap<>();
-            for (SkillFilter sf : sFilters) {
-
-                resumeIdAndCount.put(sf.getResumeId(), resumeIdAndCount.getOrDefault(sf.getResumeId(), 0) + 1);
-
-            }
-
-            Set<Integer> key = resumeIdAndCount.keySet();
-            HashMap<Integer, Integer> resumeIdAndCount2 = new HashMap<>();
-            for (Integer k : key) {
-                Integer count = resumeIdAndCount.getOrDefault(k, 0);
-
-                if (count >= 2) {
-                    resumeIdAndCount2.put(k, count);
-                }
-
-            }
-
-            // 내림차순 정렬
-            List<Entry<Integer, Integer>> resumeIdList = new ArrayList<>(resumeIdAndCount2.entrySet());
-            Collections.sort(resumeIdList, new Comparator<Entry<Integer, Integer>>() {
-                public int compare(Entry<Integer, Integer> c1, Entry<Integer, Integer> c2) {
-                    return c2.getValue().compareTo(c1.getValue());
-                }
-            });
-
-            // RESUME LIST
-            List<ResumeRecommendScrapDto> resumeList = new ArrayList<>();
-            for (Entry<Integer, Integer> entry : resumeIdList) {
-                ResumeRecommendDto resumePS = resumeRepository.findNameAndTitleAndSkills(entry.getKey());
-
-                ResumeRecommendScrapDto resumeScrapDto = new ResumeRecommendScrapDto(resumePS);
-
-                CompanyScrap cs = companyScrapRepository.findByCInfoIdAndResumeId(principal.getCInfoId(),
-                        resumePS.getId());
-
-                if (cs == null) {
-                    resumeScrapDto.setScrap(0);
-                } else {
-                    resumeScrapDto.setScrap(1);
-                }
-
-                resumeList.add(resumeScrapDto);
-
-            }
-
-            String title = postRepository.findById(p.getPostId()).getTitle();
-
-            ResumeWithPostInfoRecommendDto resumeAndPost = new ResumeWithPostInfoRecommendDto(p.getPostId(), title,
-                    resumeList);
-
-            resumeAndPostInfo.add(resumeAndPost);
-        }
-
-        return new ResponseEntity<>(new ResponseDto<>(1, "기업 인재 추천", resumeAndPostInfo), HttpStatus.OK);
+        List<ResumeWithPostInfoRecommendDto> postWithReumseDto = companyService.recommend();
+        return new ResponseEntity<>(new ResponseDto<>(1, "기업 인재 추천", postWithReumseDto), HttpStatus.OK);
     }
 
 }
